Added RestController and returned the JSON
