+++ src/main/resources/log4j.xml
@@ -5,7 +5,7 @@
 	<appender name="console" class="org.apache.log4j.ConsoleAppender">
 		<param name="Target" value="System.out" />
 		<layout class="org.apache.log4j.PatternLayout">
-			<param name="ConversionPattern" value="[%p] %d [%C{1}:%L] - %m%n" />
+			<param name="ConversionPattern" value="%d{MM-dd-yyyy HH:mm:ss,SSS} %p [%C{1}:%L] - %m%n" />
 		</layout>
 	</appender>
